+++ packages/cfd/src/Containers/jurisdiction-modal-content.tsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import RootStore from 'Stores/index';
 import { connect } from 'Stores/connect';
-import { Icon, Text, Popover } from '@deriv/components';
-import { Localize, localize } from '@deriv/translations';
-import classNames from 'classnames';
-import { jurisdiction_contents } from 'Constants/jurisdiction-contents';
+import { Text } from '@deriv/components';
+import { Localize } from '@deriv/translations';
 import CfdCheckBoxForAccounts from './jurisdiction-modal-content/cfd-checkbox-for-accounts';
 import { TExistingData, TAvailableAccountAPI } from 'Components/props.types';
+import JurisdictionCard from './jurisdiction-modal-content/jurisdiction-card';
 
 type TJurisdictionModalContent = {
     account_type: string;
@@ -16,7 +15,6 @@ type TJurisdictionModalContent = {
     financial_available_accounts: TAvailableAccountAPI;
     poa_status: string;
     poi_status: string;
-    is_eu: boolean;
     is_fully_authenticated: boolean;
     poi_poa_pending: boolean;
     checked: boolean;
@@ -31,35 +29,10 @@ type TJurisdictionModalContent = {
     poa_verified: boolean;
     poi_acknowledged_for_bvi_labuan_maltainvest: boolean;
     need_poi_for_vanuatu: boolean;
-    need_poi_for_bvi_labuan: boolean;
+    need_poi_for_bvi_labuan_maltainvest: boolean;
     poi_acknowledged_for_vanuatu: boolean;
 };
 
-type TJurisdictionCard = {
-    jurisdiction_selected_shortcode: string;
-    synthetic_available_accounts: TAvailableAccountAPI;
-    financial_available_accounts: TAvailableAccountAPI;
-    account_type: string;
-    poa_status: string;
-    poi_status: string;
-    poi_poa_none: boolean;
-    setJurisdictionSelectedShortcode: (card_type: string) => void;
-    type_of_card: string;
-    disabled: boolean;
-    poa_failed: boolean;
-    poi_failed: boolean;
-    poa_acknowledged: boolean;
-    poi_acknowledged: boolean;
-    is_fully_authenticated: boolean;
-    is_virtual: boolean;
-    poi_verified_for_vanuatu: boolean;
-    poi_verified_for_bvi_labuan_maltainvest: boolean;
-    poa_verified: boolean;
-    poi_acknowledged_for_bvi_labuan_maltainvest: boolean;
-    need_poi_for_vanuatu: boolean;
-    need_poi_for_bvi_labuan: boolean;
-    poi_acknowledged_for_vanuatu: boolean;
-};
 const StatusCodes = {
     none: 'none',
     pending: 'pending',
@@ -69,51 +42,63 @@ const StatusCodes = {
     suspected: 'suspected',
 };
 
-const JurisdictionCard = ({
+const JurisdictionModalContent = ({
     jurisdiction_selected_shortcode,
+    account_type,
+    setJurisdictionSelectedShortcode,
     synthetic_available_accounts,
     financial_available_accounts,
-    account_type,
-    poi_poa_none,
-    setJurisdictionSelectedShortcode,
-    type_of_card,
-    disabled,
-    poa_failed,
-    poa_acknowledged,
-    poi_acknowledged,
+    poa_status,
+    poi_status,
+
     is_fully_authenticated,
+    checked,
+    setChecked,
+    real_synthetic_accounts_existing_data,
+    real_financial_accounts_existing_data,
+    poa_failed,
+    poi_failed,
     is_virtual,
     poi_verified_for_vanuatu,
     poi_verified_for_bvi_labuan_maltainvest,
     poa_verified,
     poi_acknowledged_for_bvi_labuan_maltainvest,
     need_poi_for_vanuatu,
-    need_poi_for_bvi_labuan,
+    need_poi_for_bvi_labuan_maltainvest,
     poi_acknowledged_for_vanuatu,
-}: TJurisdictionCard) => {
+}: TJurisdictionModalContent) => {
     const card_classname = `cfd-jurisdiction-card--${account_type}`;
-    const number_of_synthetic_accounts_to_be_shown = synthetic_available_accounts?.length;
-    const number_of_financial_accounts_to_be_shown = financial_available_accounts?.length;
 
-    const [number_of_cards] = React.useState(
-        account_type === 'synthetic'
-            ? number_of_synthetic_accounts_to_be_shown
-            : number_of_financial_accounts_to_be_shown
-    );
+    const poa_none = poa_status === StatusCodes.none;
+    const poi_none = poi_status === StatusCodes.none;
+    const poi_poa_none = poi_none || poa_none;
 
-    const cardSelection = (cardType: string) => {
-        if (jurisdiction_selected_shortcode === cardType) {
-            setJurisdictionSelectedShortcode('');
-        } else {
-            setJurisdictionSelectedShortcode(cardType);
-        }
+    const poa_acknowledged = poa_status === StatusCodes.pending || poa_status === StatusCodes.verified;
+    const poi_acknowledged = poi_status === StatusCodes.pending || poi_status === StatusCodes.verified;
+
+    const poi_poa_verified = poi_status === StatusCodes.verified && poa_status === StatusCodes.verified;
+
+    const cardsToBeShown = (type_of_card: string) => {
+        const is_available =
+            account_type === 'synthetic'
+                ? synthetic_available_accounts?.some(account => account.shortcode === type_of_card)
+                : financial_available_accounts?.some(account => account.shortcode === type_of_card);
+        return is_available;
     };
 
-    const Checkmark = () => (
-        <Icon icon='IcCheckmark' className={`${card_classname}__bullet-wrapper--checkmark`} color={'green'} size={18} />
-    );
+    const disableCard = (type_of_card: string) => {
+        if (is_virtual && type_of_card !== 'svg') {
+            return true;
+        }
+        const is_available =
+            account_type === 'synthetic'
+                ? real_synthetic_accounts_existing_data?.some(account => account.landing_company_short === type_of_card)
+                : real_financial_accounts_existing_data?.some(
+                      account => account.landing_company_short === type_of_card
+                  );
 
-    const OneOrTwoCards = number_of_cards === 1 || number_of_cards === 2;
+        return is_available;
+    };
 
     const getAccountTitle = () => {
         switch (account_type) {
@@ -126,7 +111,7 @@ const JurisdictionCard = ({
         }
     };
 
-    const getTypeTitle = () => {
+    const getTypeTitle = (type_of_card: string) => {
         switch (type_of_card) {
             case 'bvi':
                 return 'BVI';
@@ -138,7 +123,8 @@ const JurisdictionCard = ({
                 return '';
         }
     };
-    const VerificationStatuses = () => {
+
+    const VerificationStatuses = (type_of_card: string) => {
         if (is_virtual && type_of_card !== 'svg') {
             return (
                 <div className={`${card_classname}__footer--none`}>
@@ -147,14 +133,14 @@ const JurisdictionCard = ({
                             i18n_default_text='Switch to your real account to create a DMT5 {{account_title}} {{type_title}} account.'
                             values={{
                                 account_title: getAccountTitle(),
-                                type_title: getTypeTitle(),
+                                type_title: getTypeTitle(type_of_card),
                             }}
                         />
                     </Text>
                 </div>
             );
         }
-        if (!disabled && type_of_card) {
+        if (!disableCard(type_of_card) && type_of_card) {
             // account not added
             if (type_of_card === 'svg') {
                 if (!is_fully_authenticated)
@@ -239,7 +225,7 @@ const JurisdictionCard = ({
                             </div>
                         </div>
                     );
-                } else if (need_poi_for_bvi_labuan && poa_acknowledged) {
+                } else if (need_poi_for_bvi_labuan_maltainvest && poa_acknowledged) {
                     return (
                         <div className={`${card_classname}__verification-status`}>
                             <div className={`${card_classname}__verification-status--POA_POI`}>
@@ -249,7 +235,7 @@ const JurisdictionCard = ({
                             </div>
                         </div>
                     );
-                } else if (need_poi_for_bvi_labuan && !poa_acknowledged) {
+                } else if (need_poi_for_bvi_labuan_maltainvest && !poa_acknowledged) {
                     return (
                         <div className={`${card_classname}__verification-status`}>
                             <div className={`${card_classname}__verification-status--POA_POI`}>
@@ -273,140 +259,6 @@ const JurisdictionCard = ({
                 </div>
             </div>
         );
-    };
-
-    return (
-        <>
-            <div
-                className={classNames(card_classname, {
-                    [`${card_classname}--selected`]: jurisdiction_selected_shortcode === type_of_card,
-                })}
-                onClick={disabled ? () => undefined : () => cardSelection(`${type_of_card}`)}
-                style={OneOrTwoCards ? { width: '32em' } : { width: '27.6em' }}
-            >
-                {jurisdiction_contents[type_of_card as keyof typeof jurisdiction_contents].is_over_header_available && (
-                    <div className={classNames(`${card_classname}__over-header`)}>
-                        <Text as='p' color={'info-blue'} line_height='xxl' weight='bold'>
-                            <Localize
-                                i18n_default_text={
-                                    jurisdiction_contents[type_of_card as keyof typeof jurisdiction_contents]
-                                        .over_header
-                                }
-                            />
-                        </Text>
-                    </div>
-                )}
-                <div className={`${card_classname}__info-container`}>
-                    <Text as='p' color={'prominent'} weight='bold' size='sm' className={`${card_classname}__h2-header`}>
-                        <Localize
-                            i18n_default_text={
-                                jurisdiction_contents[type_of_card as keyof typeof jurisdiction_contents].header
-                            }
-                        />
-                    </Text>
-                    {account_type === 'synthetic'
-                        ? jurisdiction_contents[
-                              type_of_card as keyof typeof jurisdiction_contents
-                          ].synthetic_contents.map((item, index) => (
-                              <div className={`${card_classname}__bullet-wrapper`} key={index}>
-                                  <div>
-                                      <Checkmark />
-                                  </div>
-                                  <Text as='p' size='xs' color={'prominent'}>
-                                      <Localize i18n_default_text={item} />
-                                  </Text>
-                              </div>
-                          ))
-                        : jurisdiction_contents[
-                              type_of_card as keyof typeof jurisdiction_contents
-                          ].financial_contents.map((item, index) => (
-                              <div className={`${card_classname}__bullet-wrapper`} key={index}>
-                                  <div>
-                                      <Checkmark />
-                                  </div>
-                                  <Text as='p' size='xs' color={'prominent'}>
-                                      <Localize i18n_default_text={item} />
-                                  </Text>
-                                  {/* TODO: find a better solution */}
-                                  {/Straight-through processing/.test(item) && (
-                                      <Popover
-                                          alignment='left'
-                                          className='cfd-compare-accounts-tooltip'
-                                          classNameBubble='cfd-compare-accounts-tooltip--msg'
-                                          icon='info'
-                                          disable_message_icon
-                                          is_bubble_hover_enabled
-                                          message={localize(
-                                              'Choosing this jurisdiction will give you a Financial STP account. Your trades will go directly to the market and have tighter spreads.'
-                                          )}
-                                          zIndex={9999}
-                                      />
-                                  )}
-                              </div>
-                          ))}
-                </div>
-                <VerificationStatuses />
-            </div>
-        </>
-    );
-};
-
-const JurisdictionModalContent = ({
-    jurisdiction_selected_shortcode,
-    account_type,
-    setJurisdictionSelectedShortcode,
-    synthetic_available_accounts,
-    financial_available_accounts,
-    poa_status,
-    poi_status,
-    is_eu,
-    is_fully_authenticated,
-    checked,
-    setChecked,
-    real_synthetic_accounts_existing_data,
-    real_financial_accounts_existing_data,
-    poa_failed,
-    poi_failed,
-    is_virtual,
-    poi_verified_for_vanuatu,
-    poi_verified_for_bvi_labuan_maltainvest,
-    poa_verified,
-    poi_acknowledged_for_bvi_labuan_maltainvest,
-    need_poi_for_vanuatu,
-    need_poi_for_bvi_labuan,
-    poi_acknowledged_for_vanuatu,
-}: TJurisdictionModalContent) => {
-    const card_classname = `cfd-jurisdiction-card--${account_type}`;
-
-    const poa_none = poa_status === StatusCodes.none;
-    const poi_none = poi_status === StatusCodes.none;
-    const poi_poa_none = poi_none || poa_none;
-
-    const poa_acknowledged = poa_status === StatusCodes.pending || poa_status === StatusCodes.verified;
-    const poi_acknowledged = poi_status === StatusCodes.pending || poi_status === StatusCodes.verified;
-
-    const poi_poa_verified = poi_status === StatusCodes.verified && poa_status === StatusCodes.verified;
-
-    const cardsToBeShown = (type_of_card: string) => {
-        const is_available =
-            account_type === 'synthetic'
-                ? synthetic_available_accounts?.some(account => account.shortcode === type_of_card)
-                : financial_available_accounts?.some(account => account.shortcode === type_of_card);
-        return is_available;
-    };
-
-    const disableCard = (type_of_card: string) => {
-        if (is_virtual && type_of_card !== 'svg') {
-            return true;
-        }
-        const is_available =
-            account_type === 'synthetic'
-                ? real_synthetic_accounts_existing_data?.some(account => account.landing_company_short === type_of_card)
-                : real_financial_accounts_existing_data?.some(
-                      account => account.landing_company_short === type_of_card
-                  );
-
-        return is_available;
     };
 
     const ModalFootNote = () => {
@@ -502,7 +354,7 @@ const JurisdictionModalContent = ({
                     (jurisdiction_selected_shortcode === 'bvi' ||
                         jurisdiction_selected_shortcode === 'labuan' ||
                         jurisdiction_selected_shortcode === 'maltainvest') &&
-                    need_poi_for_bvi_labuan &&
+                    need_poi_for_bvi_labuan_maltainvest &&
                     poa_acknowledged &&
                     !poi_poa_none && (
                         <Text
@@ -560,7 +412,7 @@ const JurisdictionModalContent = ({
                     (jurisdiction_selected_shortcode === 'bvi' ||
                         jurisdiction_selected_shortcode === 'labuan' ||
                         jurisdiction_selected_shortcode === 'maltainvest') &&
-                    need_poi_for_bvi_labuan && (
+                    need_poi_for_bvi_labuan_maltainvest && (
                         <Text
                             as='p'
                             color='prominent'
@@ -616,25 +468,8 @@ const JurisdictionModalContent = ({
                                 synthetic_available_accounts={synthetic_available_accounts}
                                 financial_available_accounts={financial_available_accounts}
                                 account_type={account_type}
-                                poa_status={poa_status}
-                                poi_status={poi_status}
                                 setJurisdictionSelectedShortcode={setJurisdictionSelectedShortcode}
-                                poa_failed={poa_failed}
-                                poi_failed={poi_failed}
-                                poi_poa_none={poi_poa_none}
-                                poa_acknowledged={poa_acknowledged}
-                                poi_acknowledged={poi_acknowledged}
-                                is_fully_authenticated={is_fully_authenticated}
-                                is_virtual={is_virtual}
-                                poi_verified_for_vanuatu={poi_verified_for_vanuatu}
-                                poi_verified_for_bvi_labuan_maltainvest={poi_verified_for_bvi_labuan_maltainvest}
-                                poa_verified={poa_verified}
-                                poi_acknowledged_for_bvi_labuan_maltainvest={
-                                    poi_acknowledged_for_bvi_labuan_maltainvest
-                                }
-                                need_poi_for_vanuatu={need_poi_for_vanuatu}
-                                need_poi_for_bvi_labuan={need_poi_for_bvi_labuan}
-                                poi_acknowledged_for_vanuatu={poi_acknowledged_for_vanuatu}
+                                banner={VerificationStatuses(card)}
                             />
                         )
                 )}
@@ -648,117 +483,6 @@ const JurisdictionModalContent = ({
         </>
     );
 };
-
-// {cardsToBeShown('maltainvest') && is_eu && (
-//     <JurisdictionCard
-//         type_of_card='maltainvest'
-//         jurisdiction_selected_shortcode={jurisdiction_selected_shortcode}
-//         synthetic_available_accounts={synthetic_available_accounts}
-//         financial_available_accounts={financial_available_accounts}
-//         account_type={account_type}
-//         poa_status={poa_status}
-//         poi_status={poi_status}
-//         setJurisdictionSelectedShortcode={setJurisdictionSelectedShortcode}
-//         disabled={disableCard('maltainvest')}
-//         poa_failed={poa_failed}
-//         poi_failed={poi_failed}
-//         poi_poa_none={poi_poa_none}
-//         poa_acknowledged={poa_acknowledged}
-//         poi_acknowledged={poi_acknowledged}
-//         is_fully_authenticated={is_fully_authenticated}
-//         is_virtual={is_virtual}
-//         poi_verified_for_vanuatu={poi_verified_for_vanuatu}
-//         poi_verified_for_bvi_labuan_maltainvest={poi_verified_for_bvi_labuan_maltainvest}
-//         poa_verified={poa_verified}
-//         poi_acknowledged_for_bvi_labuan_maltainvest={poi_acknowledged_for_bvi_labuan_maltainvest}
-//         need_poi_for_vanuatu={need_poi_for_vanuatu}
-//         need_poi_for_bvi_labuan={need_poi_for_bvi_labuan}
-//         poi_acknowledged_for_vanuatu={poi_acknowledged_for_vanuatu}
-//     />
-// )}
-
-// {cardsToBeShown('vanuatu') && (
-//     <JurisdictionCard
-//         type_of_card='vanuatu'
-//         jurisdiction_selected_shortcode={jurisdiction_selected_shortcode}
-//         synthetic_available_accounts={synthetic_available_accounts}
-//         financial_available_accounts={financial_available_accounts}
-//         account_type={account_type}
-//         poa_status={poa_status}
-//         poi_status={poi_status}
-//         setJurisdictionSelectedShortcode={setJurisdictionSelectedShortcode}
-//         disabled={disableCard('vanuatu')}
-//         poa_failed={poa_failed}
-//         poi_failed={poi_failed}
-//         poi_poa_none={poi_poa_none}
-//         poa_acknowledged={poa_acknowledged}
-//         poi_acknowledged={poi_acknowledged}
-//         is_fully_authenticated={is_fully_authenticated}
-//         is_virtual={is_virtual}
-//         poi_verified_for_vanuatu={poi_verified_for_vanuatu}
-//         poi_verified_for_bvi_labuan_maltainvest={poi_verified_for_bvi_labuan_maltainvest}
-//         poa_verified={poa_verified}
-//         poi_acknowledged_for_bvi_labuan_maltainvest={poi_acknowledged_for_bvi_labuan_maltainvest}
-//         need_poi_for_vanuatu={need_poi_for_vanuatu}
-//         need_poi_for_bvi_labuan={need_poi_for_bvi_labuan}
-//         poi_acknowledged_for_vanuatu={poi_acknowledged_for_vanuatu}
-//     />
-// )}
-// {cardsToBeShown('labuan') && (
-//     <JurisdictionCard
-//         type_of_card='labuan'
-//         jurisdiction_selected_shortcode={jurisdiction_selected_shortcode}
-//         synthetic_available_accounts={synthetic_available_accounts}
-//         financial_available_accounts={financial_available_accounts}
-//         account_type={account_type}
-//         poa_status={poa_status}
-//         poi_status={poi_status}
-//         setJurisdictionSelectedShortcode={setJurisdictionSelectedShortcode}
-//         disabled={disableCard('labuan')}
-//         poa_failed={poa_failed}
-//         poi_failed={poi_failed}
-//         poi_poa_none={poi_poa_none}
-//         poa_acknowledged={poa_acknowledged}
-//         poi_acknowledged={poi_acknowledged}
-//         is_fully_authenticated={is_fully_authenticated}
-//         is_virtual={is_virtual}
-//         poi_verified_for_vanuatu={poi_verified_for_vanuatu}
-//         poi_verified_for_bvi_labuan_maltainvest={poi_verified_for_bvi_labuan_maltainvest}
-//         poa_verified={poa_verified}
-//         poi_acknowledged_for_bvi_labuan_maltainvest={poi_acknowledged_for_bvi_labuan_maltainvest}
-//         need_poi_for_vanuatu={need_poi_for_vanuatu}
-//         need_poi_for_bvi_labuan={need_poi_for_bvi_labuan}
-//         poi_acknowledged_for_vanuatu={poi_acknowledged_for_vanuatu}
-//     />
-// )}
-
-// {cardsToBeShown('svg') && (
-//     <JurisdictionCard
-//         type_of_card='svg'
-//         jurisdiction_selected_shortcode={jurisdiction_selected_shortcode}
-//         synthetic_available_accounts={synthetic_available_accounts}
-//         financial_available_accounts={financial_available_accounts}
-//         account_type={account_type}
-//         poa_status={poa_status}
-//         poi_status={poi_status}
-//         setJurisdictionSelectedShortcode={setJurisdictionSelectedShortcode}
-//         disabled={disableCard('svg')}
-//         poa_failed={poa_failed}
-//         poi_failed={poi_failed}
-//         poi_poa_none={poi_poa_none}
-//         poa_acknowledged={poa_acknowledged}
-//         poi_acknowledged={poi_acknowledged}
-//         is_fully_authenticated={is_fully_authenticated}
-//         is_virtual={is_virtual}
-//         poi_verified_for_vanuatu={poi_verified_for_vanuatu}
-//         poi_verified_for_bvi_labuan_maltainvest={poi_verified_for_bvi_labuan_maltainvest}
-//         poa_verified={poa_verified}
-//         poi_acknowledged_for_bvi_labuan_maltainvest={poi_acknowledged_for_bvi_labuan_maltainvest}
-//         need_poi_for_vanuatu={need_poi_for_vanuatu}
-//         need_poi_for_bvi_labuan={need_poi_for_bvi_labuan}
-//         poi_acknowledged_for_vanuatu={poi_acknowledged_for_vanuatu}
-//     />
-// )}
 
 export default connect(({ modules: { cfd }, client }: RootStore) => ({
     account_status: client.account_status,
